perf: split time entry wire types
